base: Increasing the total number of possible ranges in a system.

+++ src/python/pybind11/core.cc
@@ -147,7 +147,7 @@ init_range(py::module_ &m_native)
     py::class_<AddrRange>(m, "AddrRange")
         .def(py::init<>())
         .def(py::init<Addr &, Addr &>())
-        .def(py::init<Addr, Addr, const std::vector<Addr> &, uint8_t>())
+        .def(py::init<Addr, Addr, const std::vector<Addr> &, uint32_t>())
         .def(py::init<const std::vector<AddrRange> &>())
         .def(py::init<Addr, Addr, uint8_t, uint8_t, uint8_t, uint8_t>())
